backend/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==5.0.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
multidict==6.0.5
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
twilio==9.0.5
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
zope.interface==6.3
